Move methods shared by deploy_egg and deploy_reqs to utils
To add --from-pypi to deploy_egg it would need some functions that were
already on deploy_reqs, which would lead us to a dependency cycle.
